pyproject.toml
[tool.poetry]
name = "books-base"
version = "0.1.0"
description = ""
authors = ["sonic_72_kon <kko160305@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.13.1"
environs = "^11.0.0"
redis = "^5.2.0"
betterlogging = "^1.0.0"
greenlet = "^3.1.1"
backoff = "^2.2.1"
ujson = "^5.10.0"
fluent-runtime = "^0.4.0"
yoomoney = "^0.1.1"
requests = "^2.32.3"
pydantic-settings = "^2.6.0"
apscheduler = "^3.10.4"
fastapi = "^0.115.3"
uvicorn = {version = "^0.32.0", extras = ["standard"]}
alembic = "^1.13.3"
sqlalchemy = {version = "^2.0.36", extras = ["asyncio"]}
asyncpg = "^0.30.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"
aiosqlite = "^0.20.0"
orjson = "^3.10.10"
rapidfuzz = "^3.10.1"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
